+++ src/Commerce/Gateways/Omnipay/Buckaroo_CreditCard_GatewayAdapter.php
@@ -1,0 +1,10 @@
+<?php
+namespace Commerce\Gateways\Omnipay;
+
+class Buckaroo_CreditCard_GatewayAdapter extends \Commerce\Gateways\OffsiteGatewayAdapter
+{
+    public function handle()
+    {
+        return 'Buckaroo_CreditCard';
+    }
+}